Added homework_1.rb
This program asks for a city name, then displays the correponding area code. It can be fixed up quite a bit more (like being able to take capital letters as well, or even accepting y to continue), but since there are more projects to tackle, this will do just fine.