Started Alpha Blog Project
This project is supposed to be made with plain HTML and CSS, but I feel I've done enough of that to just make a plain old site. Therefore, I've imported Bootstrap 4 because I want to play with their new framework, and I can instead focus on making the Rails application run rather than trying to style everything.