Added Article to db.
Need to cement naming conventions with Rails so here goes: The class is capitalized and singular (ie. Article). The model file (which holds the class) is singular and lowercase (ie. article.rb). The controller file is plural and snake case (ie. article_controller.rb). Lastly, the controller class is plural and camel case (ie. ArticlesController).