Connected new article categories to db.
It wasn't making the association with :category_ids and inserting into the proper categories to the db. In order to fix, I had to use f.collection_select rather than collection_select in the form. I'm assuming that it's because collection_select by itself doesn't use the form helper, which knows the name of the model I want to put the information into. It twerks meow, so drinks all around.